Change AirBeamSessionCreator to AirBeamFixed session creator
